+++ joblib/externals/loky/backend/fork_exec.py
@@ -33,11 +33,16 @@ def close_fds(keep_fds):  # pragma: no cover
             pass
 
 
-def fork_exec(cmd, keep_fds):
+def fork_exec(cmd, keep_fds, env=None):
+
+    # copy the environment variables to set in the child process
+    env = {} if env is None else env
+    child_env = os.environ.copy()
+    child_env.update(env)
 
     pid = os.fork()
     if pid == 0:  # pragma: no cover
         close_fds(keep_fds)
-        os.execv(sys.executable, cmd)
+        os.execve(sys.executable, cmd, child_env)
     else:
         return pid